refactor(database): label select component
