+++ packages/affine/data-view/src/core/component/index.ts
@@ -1,0 +1,3 @@
+export * from './button/button.js';
+export * from './overflow/overflow.js';
+export * from './tags/index.js';
