+++ packages/affine/data-view/src/core/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './tags/index.js';
 export * from './uni-component/index.js';
 export * from './uni-icon.js';
