+++ packages/affine/data-view/src/property-presets/multi-select/cell-renderer.ts
@@ -1,12 +1,13 @@
 import { popupTargetFromElement } from '@blocksuite/affine-components/context-menu';
+import { computed } from '@preact/signals-core';
 import { html } from 'lit/static-html.js';
 
 import type { SelectTag } from '../../core/index.js';
 import type { SelectPropertyData } from '../select/define.js';
 
+import { popTagSelect } from '../../core/component/tags/multi-tag-select.js';
 import { BaseCellRenderer } from '../../core/property/index.js';
 import { createFromBaseCellRenderer } from '../../core/property/renderer.js';
-import { popTagSelect } from '../../core/utils/tags/multi-tag-select.js';
 import { createIcon } from '../../core/utils/uni-icon.js';
 import { multiSelectPropertyModelConfig } from './define.js';
 
@@ -35,7 +36,7 @@ export class MultiSelectCellEditing extends BaseCellRenderer<
           this.querySelector('affine-multi-tag-view') ?? this
         ),
         {
-          options: this._options,
+          options: this.options$,
           onOptionsChange: this._onOptionsChange,
           value: this._value,
           onChange: this._onChange,
@@ -63,9 +64,9 @@ export class MultiSelectCellEditing extends BaseCellRenderer<
     });
   };
 
-  get _options(): SelectTag[] {
+  options$ = computed(() => {
     return this.property.data$.value.options;
-  }
+  });
 
   get _value() {
     return this.value ?? [];
@@ -79,7 +80,7 @@ export class MultiSelectCellEditing extends BaseCellRenderer<
     return html`
       <affine-multi-tag-view
         .value="${this._value}"
-        .options="${this._options}"
+        .options="${this.options$.value}"
       ></affine-multi-tag-view>
     `;
   }
